+++ src/test/java/AccordingTest.java
@@ -6,8 +6,6 @@ import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
-import pageobject.Constans;
-import pageobject.MainPage;
 
 import java.util.concurrent.TimeUnit;
