--- src/test/java/pageobject/Constans.java
@@ -1,12 +1,0 @@
-package pageobject;
-
-public class Constans {
-    public static final String ACCORDING_TEXT_FOR_CHECK_0 = "Сутки — 400 рублей. Оплата курьеру — наличными или картой.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_1 = "Пока что у нас так: один заказ — один самокат. Если хотите покататься с друзьями, можете просто сделать несколько заказов — один за другим.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_2 = "Допустим, вы оформляете заказ на 8 мая. Мы привозим самокат 8 мая в течение дня. Отсчёт времени аренды начинается с момента, когда вы оплатите заказ курьеру. Если мы привезли самокат 8 мая в 20:30, суточная аренда закончится 9 мая в 20:30.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_3 = "Только начиная с завтрашнего дня. Но скоро станем расторопнее.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_4 = "Пока что нет! Но если что-то срочное — всегда можно позвонить в поддержку по красивому номеру 1010.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_5 = "Самокат приезжает к вам с полной зарядкой. Этого хватает на восемь суток — даже если будете кататься без передышек и во сне. Зарядка не понадобится.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_6 = "Да, пока самокат не привезли. Штрафа не будет, объяснительной записки тоже не попросим. Все же свои.";
-    public static final String ACCORDING_TEXT_FOR_CHECK_7 = "Да, обязательно. Всем самокатов! И Москве, и Московской области.";
-}
